merkledag: split off node.go
